Added comments, and made one more minor adjustment.
