+++ Samples/CrossUnityDependencies.Unity/Assets/CrossUnityDependencies.Unity.msb4u.sln
@@ -419,14 +419,14 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{C47857AB-4B18-4AED-94F8-E6ACB45B6484} = {DE0C24FB-D205-472C-9A12-A6889F85118C}
-		{58E82562-D72E-4510-BB63-C1DDBA4E3433} = {DE0C24FB-D205-472C-9A12-A6889F85118C}
-		{6546D776-5B41-65B4-0850-B3667F981C26} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
-		{6055BE8E-BEFD-69E4-8B49-212B09B47B2F} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
-		{15A615C7-33AA-2409-09FE-0B28B0D5143C} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
-		{645165C8-1694-74BF-BBEB-8FB0BCFD26F5} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
-		{0FDA7EBE-61AB-2164-383D-10E32EFB9C6E} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
 		{0DF1F49A-50F1-5354-5BF9-9E3DBB6674AD} = {53C6B76F-FA9E-4BC0-9475-9A0944C6D7F5}
+		{0FDA7EBE-61AB-2164-383D-10E32EFB9C6E} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
+		{15A615C7-33AA-2409-09FE-0B28B0D5143C} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
+		{58E82562-D72E-4510-BB63-C1DDBA4E3433} = {DE0C24FB-D205-472C-9A12-A6889F85118C}
+		{6055BE8E-BEFD-69E4-8B49-212B09B47B2F} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
+		{645165C8-1694-74BF-BBEB-8FB0BCFD26F5} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
+		{6546D776-5B41-65B4-0850-B3667F981C26} = {E6B8BA08-82D8-4266-93D4-0A9770D70A66}
+		{C47857AB-4B18-4AED-94F8-E6ACB45B6484} = {DE0C24FB-D205-472C-9A12-A6889F85118C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {79C75D6A-84D9-443A-98F8-4DE7AC4365DC}
