+++ Source/MSBuildTools.Unity/Packages/com.microsoft.msbuildforunity/Editor/ProjectGenerator/Scripts/Exporters/ProjectConfigurationPlatformMapping.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6fea1668e7c6cc549aac4fd9ff1b65b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
